backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
vanna>=0.7.0
openai>=1.12.0
chromadb>=0.4.0
pydantic>=2.0.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
groq>=0.4.0
pandas>=2.0.0
tabulate>=0.9.0
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
pdf2image>=1.16.0
openai-whisper>=20231117
python-multipart>=0.0.6
numpy>=1.24.0
easyocr>=1.7.0
